poc_clustering/requirements.txt
# PoC Clustering - Dependencies

# Embedding model
sentence-transformers>=2.5.0

# Vector similarity search
faiss-cpu>=1.7.4

# PyTorch (CPU-only for ARM)
torch>=2.0.0

# LLM for hashtag generation
openai>=1.12.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Utilities
numpy>=1.24.0
tqdm>=4.66.0
